+++ zest/releaser/utils.py
@@ -126,6 +126,9 @@ def suggest_version(
     levels=0,
     dev_marker=".dev0",
     final=False,
+    alpha=False,
+    beta=False,
+    rc=False,
 ):
     """Suggest new version.
 
@@ -135,6 +138,9 @@ def suggest_version(
     - feature: increase major version, 1.2.3 -> 1.3.
     - breaking: increase minor version, 1.2.3 -> 2 (well, 2.0)
     - final: remove a/b/rc, 6.0.0rc1 -> 6.0.0
+    - alpha: create alpha version: 1.0 -> 1.0a1
+    - beta: create beta version: 1.0a1 -> 1.0b1
+    - rc: create rc version: 1.0b1 -> 1.0rc1
     - less_zeroes: instead of 2.0.0, suggest 2.0.
       Only makes sense in combination with feature or breaking.
     - levels: number of levels to aim for.  3 would give: 1.2.3.
@@ -144,6 +150,24 @@ def suggest_version(
     if len(list(filter(None, [breaking, feature, final]))) > 1:
         print("ERROR: Only enable one option of breaking/feature/final.")
         sys.exit(1)
+    if len(list(filter(None, [final, alpha, beta, rc]))) > 1:
+        print("ERROR: Only enable one option of final/alpha/beta/rc.")
+        sys.exit(1)
+    # Save the original options, as we need to call this function again
+    # for some combinations of options.
+    orig_options = dict(
+        current=current,
+        feature=feature,
+        breaking=breaking,
+        less_zeroes=less_zeroes,
+        levels=levels,
+        dev_marker=dev_marker,
+        final=final,
+        alpha=alpha,
+        beta=beta,
+        rc=rc,
+    )
+
     dev = ""
     base_dev_marker = strip_last_number(dev_marker)
     if base_dev_marker in current:
@@ -151,6 +175,53 @@ def suggest_version(
         # Put the standard development marker back at the end.
         dev = dev_marker
         current = current[:index]
+    try:
+        parsed_version = parse_version(current)
+    except InvalidVersion:
+        logger.warning("Cannot parse version '%s'", current)
+        # Try to at least detect if there is a number at the end.
+        last_number = re.findall(r"\d+$", current)
+        if not last_number:
+            logger.warning(
+                "Version does not end with a number, so we can't "
+                "calculate a suggestion for a next version."
+            )
+            return
+        current = current[: -len(last_number[-1])]
+        last_number = str(int(last_number[-1]) + 1)
+        return current + last_number + dev
+    if (alpha or beta or rc) and (breaking or feature):
+        # Call this function once for each option.
+        # First do breaking/feature.
+        first_options = dict(orig_options)
+        first_options.update(dict(alpha=False, beta=False, rc=False))
+        intermediate_version = suggest_version(**first_options)
+        second_options = dict(orig_options)
+        # Then to alpha/beta/rc on the intermediate result.
+        second_options.update(
+            dict(feature=False, breaking=False, current=intermediate_version)
+        )
+        return suggest_version(**second_options)
+    if alpha:
+        if not parsed_version.pre:
+            # Create new alpha version.
+            return f"{current}a1{dev}"
+        return
+    if beta:
+        if not parsed_version.pre or parsed_version.pre[0] == "a":
+            return f"{parsed_version.base_version}b1{dev}"
+        return
+    if rc:
+        if not parsed_version.pre or parsed_version.pre[0] in ("a", "b"):
+            return f"{parsed_version.base_version}rc1{dev}"
+        return
+    if final:
+        if not parsed_version.pre:
+            logger.warning(
+                "Version is not a pre version, so we cannot "
+                "calculate a suggestion for the final version."
+            )
+            return
     # Split in first and last part, where last part is one integer and the
     # first part can contain more integers plus dots.
     current_split = current.split(".")
@@ -189,7 +260,6 @@ def suggest_version(
             # too tricky.
             return
         if final:
-            parsed_version = parse_version(current)
             if not parsed_version.pre:
                 logger.warning(
                     "Version is not a pre version, so we cannot "
@@ -1063,3 +1133,13 @@ def filename_from_test_dir(filename: str) -> str:
     this_dir = os.path.dirname(__file__)
     test_dir = os.path.join(this_dir, "tests")
     return os.path.join(test_dir, filename)
+
+
+def check_mutially_exclusive_options(options, *names):
+    count = 0
+    for name in names:
+        if getattr(options, name, 0):
+            count += 1
+            if count > 1:
+                print(f"ERROR: Only enable one option of {'/'.join(names)}.")
+                sys.exit(1)